Torchtune amd

+++ torchtune/training/_distributed.py
@@ -253,7 +253,7 @@ def load_from_full_model_state_dict(
         sharded_sd = {}
         for param_name, full_tensor in full_sd.items():
             sharded_meta_param = meta_sharded_sd.get(param_name)
-            full_tensor = full_tensor.to(sharded_meta_param.dtype).to(device)
+            full_tensor = full_tensor.to(sharded_meta_param.dtype, copy=True).to(device)
             if hasattr(sharded_meta_param, "_local_tensor") and isinstance(
                 sharded_meta_param._local_tensor, NF4Tensor
             ):
